+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Created All Groups widget [#1786](https://github.com/rokwire/illinois-app/issues/1786).
+- Load wellness tip color from the new day color API [#1788](https://github.com/rokwire/illinois-app/issues/1788).
 
 ## [4.0.33] - 2022-06-20
 ### Changed